+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 1.3.1: Aug 08, 2024
+* Refactor render subscribe form function ([DESCW-2487](https://citz-gdx.atlassian.net/browse/DESCW-2487))
+
 ## 1.3.0: July 24, 2024
 * Remove dismiss button from emergency alert banner ([DESCW-2512](https://citz-gdx.atlassian.net/browse/DESCW-2512))
 * Add state of emergency alert banner ([DESCW-2434](https://citz-gdx.atlassian.net/browse/DESCW-2434))